Add authentication and ownership

+++ listings/views.py
@@ -5,11 +5,14 @@ from .serializers.common import ListingSerializer
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
+
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
 
 # Create your views here.
 
 class ListingListView(APIView):
+    permission_classes = (IsAuthenticatedOrReadOnly, )
 
     # endpoint: /listings/
     # description: return all listings to the user
@@ -24,8 +27,9 @@ class ListingListView(APIView):
     # description: using request body, add new listing to listing table
     def post(self, request):
         print(request.data)
-        listing_to_add = ListingSerializer(data=request.data)
+        request.data['owner'] = request.user.id
         try:
+            listing_to_add = ListingSerializer(data=request.data)
             if listing_to_add.is_valid():
                 print(listing_to_add)
                 listing_to_add.save()
@@ -38,6 +42,7 @@ class ListingListView(APIView):
 
 
 class ListingDetailView(APIView):
+    permission_classes = (IsAuthenticatedOrReadOnly, )
 
     # custom generic get funciton for 1 listing
     def get_listing(self, pk):
@@ -59,8 +64,10 @@ class ListingDetailView(APIView):
     # endpoint: /listings/:pk
     # description: take user data in the instance and validate updates before saving
     def put(self, request, pk):
-        listing = self.get_listing(pk)
         try:
+            listing = self.get_listing(pk)
+            if listing.owner != request.user:
+                raise PermissionDenied('Unauthorized')
             serialized_listing_to_update = ListingSerializer(listing, request.data, partial=True)
             if serialized_listing_to_update.is_valid(): 
                 serialized_listing_to_update.save()
@@ -72,9 +79,11 @@ class ListingDetailView(APIView):
 
     # endpoint: /listings/:pk
     # description: delete a listing
-    def delete(self, _request, pk):
-        listing_to_delete = self.get_listing(pk)
+    def delete(self, request, pk):
         try:
+            listing_to_delete = self.get_listing(pk)
+            if listing_to_delete.owner != request.user:
+                raise PermissionDenied('Unauthorized')
             listing_to_delete.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         except Listing.DoesNotExist as e:
